fixed problem with space and abbrev.
At some point a script seems to have been run that changes the way
abbrev. element and spaces function. This resulted in spaces between
elements, which means they essentially vanish and do not display in
html. Went through a batch of these and changed so that space is now
preceding the closing of abbr. element.